Refactor log-likelihood functions to avoid code duplication
Instead of having very similar code for every state, now the cross section and
lambdas are calculated from a list of models and branching fractions that
contribute to the desired state.

This is still not ideal, because e.g. the j/psi calls, although pretty straight
forward, are very heavy. Ideally it would be possible to re-use intermediate
models. In this case this would be the chi states, which could then be passed to
the jpsi call, which would allow for a slightly smaller call site.