Implement costh ratio fit in terms of new fitter
Done to check whether the method using the hybrid scanning/fitting approach
works the same as the 5d scanning method used for the CMS analysis. Also some
preliminary work for including the costh ratios into the global fit.

NOTE: Removed the min pt/M condition in reading in the data. Since all the
currently available data are in any case pt/M > 2, this does not change anything
for the global fit.