Make it possible to restrict polarization corrections to physical range
Polarizations outside the physically allowed range are still allowed in the fit.
However, the corrections will be determined according the extreme value the
current polarization parameter is exceeding. Hence, corrections applied to the
cross section data will always be from the physically allowed domain, even if
the polarizations are not.