Reparametrize fit in terms of total and longitudinal cross sections
This is parametrization allows to do constraints on the parameters that are
closer to what is observed on data, e.g. that all total cross sections have the
same pT/M behavior.

Technically it means to more or less replace the transverse beta with the total
beta and changing how the direct cross section and the corresponding direct
polarizations are calculated.